requirements.txt
asgiref==3.6.0
Django==4.1.7
sqlparse==0.4.3
asgiref==3.6.0
cffi==1.15.1
cryptography==40.0.1
Django==4.1.7
django-rest-knox==4.2.0
djangorestframework==3.14.0
pycparser==2.21
pytz==2023.3
sqlparse==0.4.3
